# src/swagger.yaml
openapi: 3.0.3
info:
  title: MedVehicule API
  description: API REST de gestion de flotte de véhicules hospitaliers
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/auth/login:
    post:
      summary: Connexion
      tags: [Authentification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Token et utilisateur }
        '401': { description: Identifiants incorrects }
  /api/auth/me:
    get:
      summary: Utilisateur courant
      tags: [Authentification]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Utilisateur }
...
  /api/authorizations:
    get:
      summary: Liste des droits d'utilisation
      tags: [Autorisations]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Liste }
    post:
      summary: Accorder un droit
      tags: [Autorisations]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [userId, vehicleTypeId]
              properties:
                userId: { type: string }
                vehicleTypeId: { type: string }
                expiresAt: { type: string, format: date-time }
      responses:
        '201': { description: Créé }
  /api/authorizations/user/{userId}:
    get:
      summary: Droits d'un utilisateur
      tags: [Autorisations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/authorizations/{id}/revoke:
    post:
      summary: Révoquer un droit
      tags: [Autorisations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/bookings:
    get:
      summary: Liste des réservations
      tags: [Réservations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: query
          schema: { type: string }
        - name: userId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [planned, in_progress, completed, cancelled] }
      responses:
        '200': { description: Liste }
    post:
      summary: Créer une réservation
      tags: [Réservations]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId, startDate]
              properties:
                vehicleId: { type: string }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
                notes: { type: string }
      responses:
        '201': { description: Créé }
  /api/bookings/{id}:
    get:
      summary: Détail réservation
      tags: [Réservations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/bookings/{id}/start:
    patch:
      summary: Démarrer la prise
      tags: [Réservations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                odometerStart: { type: integer, minimum: 0 }
      responses:
        '200': { description: OK }
  /api/bookings/{id}/complete:
    patch:
      summary: Terminer la prise (restitution)
      tags: [Réservations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [odometerEnd]
              properties:
                odometerEnd: { type: integer, minimum: 0 }
                notes: { type: string }
      responses:
        '200': { description: OK }
  /api/bookings/{id}/cancel:
    patch:
      summary: Annuler une réservation
      tags: [Réservations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/assistance:
    get:
      summary: Liste des demandes d'assistance
      tags: [Assistance]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [ouverte, en_route, traitee] }
      responses:
        '200': { description: Liste }
    post:
      summary: Créer une demande d'assistance
      tags: [Assistance]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId, type]
              properties:
                vehicleId: { type: string }
                type: { type: string, enum: [panne, accident, depannage] }
                description: { type: string }
                location: { type: string }
      responses:
        '201': { description: Créé }
  /api/assistance/{id}:
    patch:
      summary: Mettre à jour statut assistance
      tags: [Assistance]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [ouverte, en_route, traitee] }
      responses:
        '200': { description: OK }
...